requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
google-generativeai==0.3.2
pyjwt==2.8.0
bcrypt==4.1.2
gunicorn==21.2.0
